+++ common/src/main/java/net/arna/jcraft/registry/JCommandRegistry.java
@@ -14,7 +14,6 @@ public interface JCommandRegistry {
         SetStandCommand.register(dispatcher);
         ClearStandCommand.register(dispatcher);
         SetSpecCommand.register(dispatcher);
-        MoveDataCommand.register(dispatcher);
         FrameDataCommand.register(dispatcher);
         StandSkinCommand.register(dispatcher);
         StandBlockCommand.register(dispatcher);